+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Feature
 
 - Technical User Creation - API Response Changes
+- AppSubscription, Service Subscription, Company Subscription
+  - Validation of Technical User UIs for multiple tech user support
 
 ### Change